moved reproduce_fig1d() to a separate file
